+++ cmd/bosun/conf/native/native.go
@@ -1,7 +1,6 @@
 package native
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
@@ -692,102 +691,6 @@ func (c *NativeConf) loadTemplate(s *parse.SectionNode) {
 }
 
 var lookupNotificationRE = regexp.MustCompile(`^lookup\("(.*)", "(.*)"\)$`)
-
-func (c *NativeConf) SetAlert(name, alertText string) (string, error) {
-	select {
-	case c.writeLock <- true:
-		// Got Write Lock
-	default:
-		return "", fmt.Errorf("cannot write alert, write in progress")
-	}
-	defer func() {
-		<-c.writeLock
-	}()
-	a := c.GetAlert(name)
-	var newRawConf string
-	if a == nil {
-		newRawConf = writeSection(nil, c.RawText, alertText)
-	} else {
-		newRawConf = writeSection(a.Locator, c.RawText, alertText)
-	}
-	newConf, err := NewNativeConf(c.Name, newRawConf)
-	if err != nil {
-		return "", fmt.Errorf("new config not valid: %v", err)
-	}
-	if err := c.SaveConf(newConf); err != nil {
-		return "", fmt.Errorf("couldn't save config file")
-	}
-	err = c.reload()
-	if err != nil {
-		return "", err
-	}
-	return "reloaded", nil
-}
-
-func (c *NativeConf) DeleteAlert(name string) (error) {
-	select {
-	case c.writeLock <- true:
-		// Got Write Lock
-	default:
-		return fmt.Errorf("cannot write alert, write in progress")
-	}
-	defer func() {
-		<-c.writeLock
-	}()
-	a := c.GetAlert(name)
-	if a == nil {
-		return fmt.Errorf("alert %v not found", name)
-	}
-	newRawConf := removeSection(a.Locator, c.RawText)
-	newConf, err := NewNativeConf(c.Name, newRawConf)
-	if err != nil {
-		return fmt.Errorf("new config not valid: %v", err)
-	}
-	if err := c.SaveConf(newConf); err != nil {
-		return fmt.Errorf("couldn't save config file: %v", err)
-	}
-	err = c.reload()
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func writeSection(l *conf.Locator, orginalRaw, newText string) string {
-	var newRawConf bytes.Buffer
-	if l == nil {
-		newRawConf.WriteString(orginalRaw)
-		newRawConf.WriteString("\n")
-		newRawConf.WriteString(newText)
-		return newRawConf.String()
-	}
-	newRawConf.WriteString(orginalRaw[:getLocationStart(l)])
-	newRawConf.WriteString(newText)
-	newRawConf.WriteString(orginalRaw[getLocationEnd(l):])
-	return newRawConf.String()
-}
-
-func removeSection(l *conf.Locator, orginalRaw string) string {
-	var newRawConf bytes.Buffer
-	newRawConf.WriteString(orginalRaw[:getLocationStart(l)])
-	newRawConf.WriteString(orginalRaw[getLocationEnd(l):])
-	return newRawConf.String()
-}
-
-func newLocator(start int, end int) (*conf.Locator) {
-	l := &conf.Locator{}
-	l.Location = conf.NativeLocator{start, end}
-	l.LocatorType = conf.TypeNative
-	return l
-}
-
-func getLocationStart(l *conf.Locator) int {
-	return l.Location.(conf.NativeLocator)[0]
-}
-
-func getLocationEnd(l *conf.Locator) int {
-	return l.Location.(conf.NativeLocator)[1]
-}
 
 func (c *NativeConf) loadAlert(s *parse.SectionNode) {
 	name := s.Name.Text
